refactor(agent): rename executor types to runtime in lib/
