+++ mods/agent/lib/codex-appserver-runtime.ts
@@ -1,6 +1,6 @@
 /**
- * AsyncGenerator-based executor bridging the Codex AppServer JSON-RPC protocol
- * to the unified {@link AIAgentExecuter} interface.
+ * AsyncGenerator-based runtime bridging the Codex AppServer JSON-RPC protocol
+ * to the unified {@link AgentRuntime} interface.
  *
  * Communicates with the shared {@link CodexAppServerProcess} to start/resume
  * threads, handle server requests (approval, elicitation), and map server
@@ -11,10 +11,10 @@
 
 import { AsyncQueue } from "./async-queue.ts";
 import type {
-  AIAgentExecuter,
+  AgentRuntime,
   AgentEvent,
   AgentResponse,
-} from "./ai-agent-executer.ts";
+} from "./agent-runtime.ts";
 import type {
   CodexAppServerProcess,
   ThreadHandler,
@@ -54,7 +54,7 @@ type QueueMessage =
  * an event loop that maps server notifications and requests to
  * {@link AgentEvent} values yielded through an AsyncGenerator.
  */
-export class CodexAppServerExecuter implements AIAgentExecuter {
+export class CodexAppServerRuntime implements AgentRuntime {
   private readonly persona: Persona;
   private readonly settings: AgentSettings;
   private readonly workDir: string;
@@ -175,7 +175,7 @@ export class CodexAppServerExecuter implements AIAgentExecuter {
       return sessionId;
     } catch (e) {
       console.warn(
-        `[codex-appserver-executer] Failed to resume thread ${sessionId}, starting fresh:`,
+        `[codex-appserver-runtime] Failed to resume thread ${sessionId}, starting fresh:`,
         e instanceof Error ? e.message : e,
       );
       return this.startThread();
@@ -316,7 +316,7 @@ export class CodexAppServerExecuter implements AIAgentExecuter {
     pendingRequests.delete(String(id));
   }
 
-  /** Auto-decline a request the executor does not handle. */
+  /** Auto-decline a request the runtime does not handle. */
   private autoDeclineRequest(
     id: RequestId,
     method: string,
@@ -390,7 +390,7 @@ export class CodexAppServerExecuter implements AIAgentExecuter {
     // "interrupted" — emit error event so the event loop terminates cleanly
     // even for server-initiated interruptions (e.g. context exhaustion, rate limits).
     console.warn(
-      `[codex-appserver-executer] Turn interrupted by server (status: ${status})`,
+      `[codex-appserver-runtime] Turn interrupted by server (status: ${status})`,
     );
     return [{ type: "error", message: "Turn was interrupted by the server" }];
   }
@@ -712,7 +712,7 @@ export class CodexAppServerExecuter implements AIAgentExecuter {
         .sendRequest("turn/interrupt", { threadId, turnId })
         .catch((e) => {
           console.warn(
-            "[codex-appserver-executer] Failed to interrupt turn:",
+            "[codex-appserver-runtime] Failed to interrupt turn:",
             e instanceof Error ? e.message : e,
           );
         });
